+++ nengo/builder/signal.py
@@ -52,10 +52,11 @@ class Signal(object):
             assert name
         self._name = name
 
-        if not np.isscalar(initial_value) and base is None:
+        initial_value = np.asarray(initial_value)
+        if initial_value.ndim > 0 and base is None:
             self._initial_value = np.ascontiguousarray(initial_value).view()
         else:
-            self._initial_value = np.asarray(initial_value).view()
+            self._initial_value = initial_value.view()
         self._initial_value.setflags(write=False)
 
         if base is not None: