+++ nengo/transforms.py
@@ -2,42 +2,10 @@ import numpy as np
 
 import nengo
 from nengo.base import FrozenObject
-from nengo.dists import Distribution
+from nengo.dists import Distribution, DistOrArrayParam
 from nengo.exceptions import ValidationError
-from nengo.utils.compat import range
-
-
-def get_transform(transform, shape, rng=np.random):
-    """Convenience function to sample a transform or return samples.
-
-    Use this function in situations where you accept an argument that could
-    be a transform, or could be an ``array_like`` of samples.
-
-    Parameters
-    ----------
-    transform : `.Transform` or `.Distribution` or array_like
-        Source of the transform to be returned.
-    shape : tuple of int
-        Desired shape of transform
-    rng : `numpy.random.RandomState`, optional (Default: np.random)
-        Random number generator.
-
-    Returns
-    -------
-    samples : array_like
-        Sampled array
-    """
-
-    if isinstance(transform, Transform):
-        return transform.sample(rng=rng)
-    elif isinstance(transform, Distribution):
-        assert len(shape) == 2
-        return transform.sample(*shape, rng=rng)
-    else:
-        assert (transform.shape == ()
-                or transform.shape == (shape[0],)
-                or transform.shape == shape)
-        return np.array(transform)
+from nengo.params import ShapeParam, IntParam, EnumParam, BoolParam
+from nengo.utils.compat import is_array_like
 
 
 class Transform(FrozenObject):
@@ -69,6 +37,88 @@ class Transform(FrozenObject):
         raise NotImplementedError()
 
 
+class ChannelShapeParam(ShapeParam):
+    """A parameter where the value must be a shape with channels."""
+
+    def coerce(self, transform, shape):
+        if isinstance(shape, ChannelShape):
+            if shape.channels_last != transform.channels_last:
+                raise ValidationError(
+                    "transform has channels_last=%s, but input shape has "
+                    "channels_last=%s"
+                    % (transform.channels_last, shape.channels_last),
+                    attr=self.name, obj=transform)
+            super(ChannelShapeParam, self).coerce(transform, shape.shape)
+        else:
+            super(ChannelShapeParam, self).coerce(transform, shape)
+            shape = ChannelShape(shape, channels_last=transform.channels_last)
+        return shape
+
+
+class Dense(Transform):
+    """A dense transformation between an input and output signal.
+
+    Parameters
+    ----------
+    shape : tuple of int
+        The shape of the dense matrix: ``(size_out, size_in)``.
+    init : `.Distribution` or array_like, optional (Default: 1.0)
+        A Distribution used to initialize the transform matrix, or a concrete
+        instantiation for the matrix. If the matrix is square we also allow a
+        scalar (equivalent to ``np.eye(n) * init``) or a vector (equivalent to
+        ``np.diag(init)``) to represent the matrix more compactly.
+    """
+
+    shape = ShapeParam("shape", length=2, low=1)
+    init = DistOrArrayParam("init")
+
+    def __init__(self, shape, init=1.0):
+        super(Dense, self).__init__()
+
+        self.shape = shape
+
+        if is_array_like(init):
+            init = np.asarray(init, dtype=np.float64)
+
+            # check that the shape of init is compatible with the given shape
+            # for this transform
+            expected_shape = None
+            if shape[0] != shape[1]:
+                # init must be 2D if transform is not square
+                expected_shape = shape
+            elif init.ndim == 1:
+                expected_shape = (shape[0],)
+            elif init.ndim >= 2:
+                expected_shape = shape
+
+            if expected_shape is not None and init.shape != expected_shape:
+                raise ValidationError(
+                    "Shape of initial value %s does not match expected "
+                    "shape %s" % (init.shape, expected_shape), attr="init")
+
+        self.init = init
+
+    def sample(self, rng=np.random):
+        if isinstance(self.init, Distribution):
+            return self.init.sample(*self.shape, rng=rng)
+
+        return self.init
+
+    @property
+    def init_shape(self):
+        """The shape of the initial value."""
+        return (self.shape if isinstance(self.init, Distribution)
+                else self.init.shape)
+
+    @property
+    def size_in(self):
+        return self.shape[1]
+
+    @property
+    def size_out(self):
+        return self.shape[0]
+
+
 class Convolution(Transform):
     """An N-dimensional convolutional transform.
 
@@ -78,7 +128,7 @@ class Convolution(Transform):
     ----------
     n_filters : int
         The number of convolutional filters to apply
-    input_shape : tuple of int or `.ConvShape`
+    input_shape : tuple of int or `.ChannelShape`
         Shape of the input signal to the convolution; e.g.,
         ``(height, width, channels)`` for a 2D convolution with
         ``channels_last=True``.
@@ -99,8 +149,8 @@ class Convolution(Transform):
         signal (e.g., a 28x28 image with 3 channels would have shape
         ``(28, 28, 3)``).  ``False`` means that channels are the first
         dimension (e.g., ``(3, 28, 28)``).
-    kernel : `.Distribution` or `~numpy:numpy.ndarray`, optional \
-             (Default: Uniform(-1, 1))
+    init : `.Distribution` or `~numpy:numpy.ndarray`, optional \
+           (Default: Uniform(-1, 1))
         A predefined kernel with shape
         ``kernel_size + (input_channels, n_filters)``, or a ``Distribution``
         that will be used to initialize the kernel.
@@ -111,77 +161,59 @@ class Convolution(Transform):
     than convolution (because the kernel is not flipped).
     """
 
+    n_filters = IntParam("n_filters", low=1)
+    input_shape = ChannelShapeParam("input_shape", low=1)
+    kernel_size = ShapeParam("kernel_size", low=1)
+    strides = ShapeParam("strides", low=1)
+    padding = EnumParam("padding", values=("same", "valid"))
+    channels_last = BoolParam("channels_last")
+    init = DistOrArrayParam("init")
+
     def __init__(self, n_filters, input_shape, kernel_size=(3, 3),
                  strides=(1, 1), padding="valid", channels_last=True,
-                 kernel=nengo.dists.Uniform(-1, 1)):
+                 init=nengo.dists.Uniform(-1, 1)):
         super(Convolution, self).__init__()
 
-        if isinstance(input_shape, ConvShape):
-            assert input_shape.channels_last == channels_last
-        else:
-            input_shape = ConvShape(input_shape, channels_last=channels_last)
-
         self.n_filters = n_filters
+        self.channels_last = channels_last  # must be set before input_shape
         self.input_shape = input_shape
         self.kernel_size = kernel_size
         self.strides = strides
         self.padding = padding
-        self.channels_last = channels_last
-        self.kernel = kernel
-        self.dimensions = len(kernel_size)
+        self.init = init
 
-        if len(kernel_size) != input_shape.dimensions:
+        if len(kernel_size) != self.dimensions:
             raise ValidationError(
                 "Kernel dimensions (%d) do not match input dimensions (%d)"
-                % (len(kernel_size), input_shape.dimensions),
-                attr="kernel_size")
-        if len(strides) != input_shape.dimensions:
+                % (len(kernel_size), self.dimensions), attr="kernel_size")
+        if len(strides) != self.dimensions:
             raise ValidationError(
                 "Stride dimensions (%d) do not match input dimensions (%d)"
-                % (len(strides), input_shape.dimensions), attr="strides")
-        if not isinstance(kernel, Distribution):
-            if kernel.shape != self.kernel_shape:
+                % (len(strides), self.dimensions), attr="strides")
+        if not isinstance(init, Distribution):
+            if init.shape != self.kernel_shape:
                 raise ValidationError(
                     "Kernel shape %s does not match expected shape %s"
-                    % (kernel.shape, self.kernel_shape), attr="kernel")
-
-        # compute output shape
-        output_shape = np.array(input_shape.spatial_shape, dtype=np.float64)
-        if self.padding == "valid":
-            output_shape -= kernel_size
-            output_shape += 1
-        output_shape /= strides
-        output_shape = tuple(np.ceil(output_shape).astype(np.int64))
-        output_shape = (output_shape + (n_filters,) if channels_last
-                        else (n_filters,) + output_shape)
-
-        self.output_shape = ConvShape(
-            output_shape, channels_last=channels_last)
-
-        self._paramdict = {
-            "n_filters": n_filters,
-            "input_shape": self.input_shape,
-            "kernel_size": self.kernel_size,
-            "strides": self.strides
-        }
-
-    @property
-    def kernel_shape(self):
-        return self.kernel_size + (self.input_shape.n_channels, self.n_filters)
+                    % (init.shape, self.kernel_shape), attr="init")
 
     def sample(self, rng=np.random):
-        if isinstance(self.kernel, Distribution):
+        if isinstance(self.init, Distribution):
             # we sample this way so that any variancescaling distribution based
             # on n/d is scaled appropriately
             kernel = [
-                self.kernel.sample(
+                self.init.sample(
                     self.input_shape.n_channels, self.n_filters, rng=rng)
                 for _ in range(np.prod(self.kernel_size))
             ]
             kernel = np.reshape(kernel, self.kernel_shape)
         else:
-            kernel = np.array(self.kernel)
+            kernel = np.array(self.init)
         return kernel
+
+    @property
+    def kernel_shape(self):
+        """Full shape of kernel."""
+        return self.kernel_size + (self.input_shape.n_channels, self.n_filters)
 
     @property
     def size_in(self):
@@ -191,8 +223,28 @@ class Convolution(Transform):
     def size_out(self):
         return self.output_shape.size
 
+    @property
+    def dimensions(self):
+        """Dimensionality of convolution."""
+        return self.input_shape.dimensions
 
-class ConvShape(object):
+    @property
+    def output_shape(self):
+        """Output shape after applying convolution to input."""
+        output_shape = np.array(
+            self.input_shape.spatial_shape, dtype=np.float64)
+        if self.padding == "valid":
+            output_shape -= self.kernel_size
+            output_shape += 1
+        output_shape /= self.strides
+        output_shape = tuple(np.ceil(output_shape).astype(np.int64))
+        output_shape = (output_shape + (self.n_filters,) if self.channels_last
+                        else (self.n_filters,) + output_shape)
+
+        return ChannelShape(output_shape, channels_last=self.channels_last)
+
+
+class ChannelShape(object):
     """Represents shape information with variable channel position.
 
     Parameters
@@ -204,6 +256,7 @@ class ConvShape(object):
         and the rest are spatial dimensions. Otherwise, the first item in
         ``shape`` is the channel dimension.
     """
+
     def __init__(self, shape, channels_last=True):
         self.shape = tuple(shape)
         self.channels_last = channels_last